Move things not specific to the V2 schema to the top level module
