+++ lib/nylas.rb
@@ -4,14 +4,29 @@ require 'rest-client'
 require 'ostruct'
 require 'forwardable'
 
-require_relative 'nylas/logging'
+require 'nylas/version'
 
+require_relative 'nylas/logging'
 require_relative 'nylas/registry'
 require_relative 'nylas/to_query'
-
 require_relative 'nylas/types_filter'
+require_relative 'nylas/types'
+require_relative 'nylas/constraints'
 
-require 'nylas/version'
+require_relative 'nylas/collection'
+require_relative 'nylas/model'
+
+
+
+require_relative 'nylas/email_address'
+require_relative 'nylas/im_address'
+require_relative 'nylas/physical_address'
+require_relative 'nylas/phone_number'
+require_relative 'nylas/web_page'
+require_relative 'nylas/nylas_date'
+
+
+
 
 require 'nylas/v2'
