+++ lib/nylas/v2.rb
@@ -1,17 +1,4 @@
-require_relative 'v2/types'
-require_relative 'v2/collection'
-require_relative 'v2/model'
-
-require_relative 'v2/constraints'
-
 require_relative 'v2/contact'
-
-require_relative 'v2/email_address'
-require_relative 'v2/im_address'
-require_relative 'v2/physical_address'
-require_relative 'v2/phone_number'
-require_relative 'v2/web_page'
-require_relative 'v2/nylas_date'
 
 module Nylas
   module V2
