--- lib/nylas/v2/email_address.rb
@@ -1,16 +1,0 @@
-module Nylas
-  module V2
-    class EmailAddress
-      include Model::Attributable
-      attribute :type, :string
-      attribute :email, :string
-    end
-
-    class EmailAddressType < Types::HashType
-      casts_to EmailAddress
-    end
-
-    Types.registry[:email_address] = EmailAddressType.new
-  end
-end
-
